+++ contracts/libraries/DiscountFactor.sol
@@ -10,12 +10,12 @@ library DiscountFactor {
     uint256 internal constant NON_ANNUAL_TERMS = 3;
 
     function determineDF(
-        uint256 rate, 
-        uint256 term, 
-        uint256[] memory cache, 
+        uint256 rate,
+        uint256 term,
+        uint256[] memory cache,
         uint256 dfSum,
         uint256 index
-    ) internal pure returns (uint256 df) { 
+    ) internal pure returns (uint256 df) {
         if (term < 365) {
             df = BP.mul(BP).div((BP.add(rate.mul(term).div(360))));
         } else if (term == 365) {
@@ -28,7 +28,7 @@ library DiscountFactor {
             } else {
                 df = BP.mul(BP.sub(rate.mul(dfSum).div(BP))).div(BP.add(rate));
             }
-            dfSum = dfSum.add(df);            
+            dfSum = dfSum.add(df);
         }
 
         cache[index] = df;
@@ -36,37 +36,41 @@ library DiscountFactor {
         return dfSum;
     }
 
-    function calculateDFs(
-        uint256[] memory rates, 
-        uint256[] memory terms
-    ) public pure returns (
-        uint256[] memory,
-        uint256[] memory
-    ) {
+    function calculateDFs(uint256[] memory rates, uint256[] memory terms)
+        public
+        pure
+        returns (uint256[] memory, uint256[] memory)
+    {
         require(rates.length == terms.length, "INVALID_PARAMS");
 
         (
-            uint256[] memory bootstrapedRates, 
+            uint256[] memory bootstrapedRates,
             uint256[] memory bootstrapedTerms
         ) = bootstrapTerms(rates, terms);
 
-        uint len = bootstrapedTerms.length;
+        uint256 len = bootstrapedTerms.length;
         uint256[] memory dfs = new uint256[](len);
         uint256 dfSum;
 
-        for (uint i = 0; i < len; i++) {
-            dfSum = determineDF(bootstrapedRates[i], bootstrapedTerms[i], dfs, dfSum, i);
+        for (uint256 i = 0; i < len; i++) {
+            dfSum = determineDF(
+                bootstrapedRates[i],
+                bootstrapedTerms[i],
+                dfs,
+                dfSum,
+                i
+            );
         }
 
         return (dfs, bootstrapedTerms);
     }
 
-    function maxDFs(uint maxTerm) internal pure returns(uint) { 
+    function maxDFs(uint256 maxTerm) internal pure returns (uint256) {
         return maxTerm.div(365).add(NON_ANNUAL_TERMS);
     }
 
     struct TermBootstrapingLocalVars {
-        uint extendedTerms;
+        uint256 extendedTerms;
         uint256 delta;
         uint256 numItems;
         uint256 lastKnownRate;
@@ -77,11 +81,12 @@ library DiscountFactor {
         uint256 step;
     }
 
-    function bootstrapTerms(
-        uint256[] memory rates, 
-        uint256[] memory terms
-    ) public pure returns (uint256[] memory, uint256[] memory) {
-        uint len = maxDFs(terms[terms.length - 1]);
+    function bootstrapTerms(uint256[] memory rates, uint256[] memory terms)
+        public
+        pure
+        returns (uint256[] memory, uint256[] memory)
+    {
+        uint256 len = maxDFs(terms[terms.length - 1]);
 
         uint256[] memory filledRates = new uint256[](len);
         uint256[] memory filledTerms = new uint256[](len);
@@ -96,7 +101,7 @@ library DiscountFactor {
             vars.delta = terms[i + 1].sub(terms[i]);
 
             if (vars.delta <= 365) {
-                filledRates[i] = rates[i]; 
+                filledRates[i] = rates[i];
                 filledTerms[i] = terms[i];
                 continue;
             }
@@ -110,22 +115,28 @@ library DiscountFactor {
             }
             vars.nextKnownRate = rates[i + 1];
             vars.nextKnownTerm = terms[i + 1];
-            vars.upwards = vars.nextKnownRate > vars.lastKnownRate ? true : false;
-            vars.deltaRate = vars.upwards ? vars.nextKnownRate.sub(vars.lastKnownRate) : vars.lastKnownRate.sub(vars.nextKnownRate);
+            vars.upwards = vars.nextKnownRate > vars.lastKnownRate
+                ? true
+                : false;
+            vars.deltaRate = vars.upwards
+                ? vars.nextKnownRate.sub(vars.lastKnownRate)
+                : vars.lastKnownRate.sub(vars.nextKnownRate);
             vars.step = vars.deltaRate.div(vars.numItems);
 
             for (uint256 j = 1; j < vars.numItems; j++) {
                 vars.extendedTerms = vars.extendedTerms.add(1);
 
                 uint256 newIndex = i.add(vars.extendedTerms);
-                uint256 missedRate = vars.upwards ? filledRates[newIndex.sub(1)].add(vars.step) : filledRates[newIndex.sub(1)].sub(vars.step);
+                uint256 missedRate = vars.upwards
+                    ? filledRates[newIndex.sub(1)].add(vars.step)
+                    : filledRates[newIndex.sub(1)].sub(vars.step);
                 uint256 missedTerm = terms[i].add(uint256(365).mul(j));
 
                 filledRates[newIndex] = missedRate;
                 filledTerms[newIndex] = missedTerm;
 
                 if (j == vars.numItems.sub(1)) {
-                    uint shifterIndex = newIndex.add(1);
+                    uint256 shifterIndex = newIndex.add(1);
 
                     filledRates[shifterIndex] = vars.nextKnownRate;
                     filledTerms[shifterIndex] = vars.nextKnownTerm;
@@ -146,21 +157,17 @@ library DiscountFactor {
     }
 
     /**
-    * @dev Triggers to adjust discount factors by interpolating to current loan maturity
-    * @param discountFactors Discount factors array
-    * @param terms Array of terms
-    * @param date Date to calculate discount factors for 
-    *
-    */
+     * @dev Triggers to adjust discount factors by interpolating to current loan maturity
+     * @param discountFactors Discount factors array
+     * @param terms Array of terms
+     * @param date Date to calculate discount factors for
+     *
+     */
     function interpolateDF(
-        uint256[] memory discountFactors, 
+        uint256[] memory discountFactors,
         uint256[] memory terms,
         uint256 date
-    )
-        public
-        view
-        returns (uint256)
-    {
+    ) public view returns (uint256) {
         DFInterpolationLocalVars memory vars;
         vars.timeDelta = date.sub(block.timestamp);
 
@@ -168,13 +175,18 @@ library DiscountFactor {
             vars.termSeconds = terms[0].mul(86400);
             vars.left = vars.termSeconds.sub(vars.timeDelta);
 
-            return (BP.mul(vars.left).add(discountFactors[0].mul(vars.timeDelta))).div(vars.termSeconds);
+            return
+                (BP.mul(vars.left).add(discountFactors[0].mul(vars.timeDelta)))
+                    .div(vars.termSeconds);
         } else {
             for (uint256 i = 1; i < terms.length; i++) {
                 vars.termSeconds = terms[i].mul(86400);
                 vars.prevTermSeconds = terms[i - 1].mul(86400);
 
-                if (vars.prevTermSeconds < vars.timeDelta && vars.timeDelta <= vars.termSeconds) {
+                if (
+                    vars.prevTermSeconds < vars.timeDelta &&
+                    vars.timeDelta <= vars.termSeconds
+                ) {
                     vars.left = vars.timeDelta.sub(vars.prevTermSeconds);
 
                     if (vars.left == 0) {
@@ -188,10 +200,13 @@ library DiscountFactor {
 
                     vars.total = vars.termSeconds.sub(vars.prevTermSeconds);
 
-                    return ((discountFactors[i - 1].mul(vars.right)).add((discountFactors[i].mul(vars.left))).div(vars.total));
+                    return (
+                        (discountFactors[i - 1].mul(vars.right))
+                            .add((discountFactors[i].mul(vars.left)))
+                            .div(vars.total)
+                    );
                 }
             }
         }
     }
-
 }
